[freebsd] Fix PPC definition and remove comments

+++ src/core/sys/posix/ucontext.d
@@ -1034,7 +1034,7 @@ else version (FreeBSD)
     }
     else version (PPC)
     {
-        alias long   __register_t;
+        alias uint   __register_t;
         alias uint   __uint32_t;
         alias ulong  __uint64_t;
 
@@ -1043,13 +1043,13 @@ else version (FreeBSD)
             int     mc_flags;
             enum _MC_FP_VALID = 0x01;
             enum _MC_AV_VALID = 0x02;
-            int     mc_onstack;     /* saved onstack flag */
-            int     mc_len;         /* sizeof(__mcontext) */
-            __uint64_t[32 * 2]  mc_avec;    /* vector register file */
+            int     mc_onstack;
+            int     mc_len;
+            __uint64_t[32 * 2]  mc_avec;
             __uint32_t[2]       mc_av;
             __register_t[42]    mc_frame;
             __uint64_t[33]      mc_fpreg;
-            __uint64_t[32]      mc_vsxfpreg;    /* low-order half of VSR0-31 */
+            __uint64_t[32]      mc_vsxfpreg;
         }
     }
     else version (PPC64)
@@ -1063,13 +1063,13 @@ else version (FreeBSD)
             int     mc_flags;
             enum _MC_FP_VALID = 0x01;
             enum _MC_AV_VALID = 0x02;
-            int     mc_onstack;     /* saved onstack flag */
-            int     mc_len;         /* sizeof(__mcontext) */
-            __uint64_t[32 * 2]  mc_avec;        /* vector register file */
+            int     mc_onstack;
+            int     mc_len;
+            __uint64_t[32 * 2]  mc_avec;
             __uint32_t[2]       mc_av;
             __register_t[42]    mc_frame;
             __uint64_t[33]      mc_fpreg;
-            __uint64_t[32]      mc_vsxfpreg;    /* low-order half of VSR0-31 */
+            __uint64_t[32]      mc_vsxfpreg;
         }
     }
